requirements.txt
cloudpickle==1.3.0
gym==0.15.7
matplotlib==3.5.2
numpy==1.19.5
pandas==1.1.5
pybulletgym.egg==info
ray==2.2.0
scikit_learn==1.1.0
scipy==1.4.1
seaborn==0.11.1
stable_baselines.egg==info
tensorflow==2.5.0
tensorflow_addons==0.14.0
tensorflow_gpu==2.5.0
tensorflow_probability==0.11.0
tensorflow_similarity==0.16.7
